rename guard and move to shared directory
