+++ src/app/shared/services/auth.service.spec.ts
@@ -1,6 +1,5 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-
 import { AuthService } from './auth.service';
 
 describe('AuthService', () => {
