+++ src/app/shared/services/auth.service.ts
@@ -4,7 +4,6 @@ import { Router } from '@angular/router';
 @Injectable({
   providedIn: 'root'
 })
-
 export class AuthService {
   USER_ACCESS_TOKEN = 'user_access_token';
   USER_TOKEN_TYPE = 'user_token_type';
