+++ src/app/shared/services/base.service.ts
@@ -9,7 +9,6 @@ import { Deserializer } from 'ts-jsonapi';
 @Injectable({
   providedIn: 'root'
 })
-
 export abstract class BaseService {
   DEFAULT_HEADER: object = {
     'Content-Type': 'application/json'
